Highlight own posts in timeline
An users own posts are now highlighted in a light variant of the
primary color theme and have a pointy border in the top right.
The pointy border is in the top left for other posts.
To obtain the current user the timeline query has been expanded to
include the logged in user. It would have been possible to get the user
by accessing the secure storage but this way is easier due to no weird
cases wether the user is logged in or not and we do not have to add a
user attribute to half of the widgets.